+++ psynet/modular_page.py
@@ -724,14 +724,35 @@ class TimedPushButtonControl(PushButtonControl):
         An optional list of textual labels to apply to the buttons,
         which the participant will see instead of ``choices``. Default: ``None``.
 
+    button_highlight_duration:
+        How long to highlight the button for once it has been clicked (seconds).
+        Defaults to 0.75 s.
+
     style:
         CSS styles to apply to the buttons. Default: ``"min-width: 100px; margin: 10px"``.
 
     arrange_vertically:
         Whether to arrange the buttons vertically. Default: ``True``.
+
+    **kwargs
+        Other arguments to pass to :class:`~psynet.modular_page.PushButtonControl`.
     """
 
     timed = True
+
+    def __init__(
+            self,
+            choices: List[str],
+            labels: Optional[List[str]] = None,
+            button_highlight_duration: float = 0.75,
+            **kwargs
+    ):
+        super().__init__(
+            choices=choices,
+            labels=labels,
+            **kwargs
+        )
+        self.button_highlight_duration = button_highlight_duration
 
     def format_answer(self, raw_answer, **kwargs):
         event_log = {**kwargs}["metadata"]["event_log"]